refactoring, dropped project on modules
