+++ src/app/user/about-page/about-page.module.ts
@@ -1,0 +1,14 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import {AboutPageRoutingModule} from "./about-page-routing.module";
+
+
+
+@NgModule({
+  declarations: [],
+  imports: [
+    CommonModule,
+    AboutPageRoutingModule
+  ]
+})
+export class AboutPageModule {}
