+++ src/app/user/catalog-page/catalog-page.module.ts
@@ -1,0 +1,14 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import {CatalogPageRoutingModule} from "./catalog-page-routing.module";
+
+
+
+@NgModule({
+  declarations: [],
+  imports: [
+    CommonModule,
+    CatalogPageRoutingModule
+  ]
+})
+export class CatalogPageModule { }
