--- src/app/user/components/tiles/tiles.component.html
@@ -1,8 +1,0 @@
-<div class="search__tab-menu tab-menu">
-  <a class="tab-menu__item"
-     *ngFor="let tile of state"
-     [style]="'background-image: url('+ tile.urlImage +')'">
-    <p class="tab-menu__name" [routerLink]="['/catalog-page', tile.identifier]" >{{tile.text}}</p>
-    <p class="tab-menu__more" [routerLink]="['/extra', tile.identifier]" >Подробнее</p>
-  </a>
-</div>
