+++ src/app/user/home-page/components/head-home-carousel/head-home-carousel.component.scss
