+++ src/app/user/home-page/components/search/home-search.component.html
@@ -1,0 +1,11 @@
+<div class="search__main-form">
+  <form class="search__form" action="#">
+    <div class="search__tabs">
+      <button class="search__tab">Поиск по номеру</button>
+      <button class="search__tab active">Поиск по марке</button>
+      <button class="search__tab">Поиск по названию товара</button>
+    </div>
+    <input class="search__input" type="text" placeholder="Введите марку">
+    <button class="search__submit-button blue-button rectangle-button" type="submit">Искать</button>
+  </form>
+</div>
