+++ src/app/user/home-page/components/search/home-search.component.scss
@@ -1,0 +1,50 @@
+@import "../../../../styles/vars";
+
+.search {
+  margin-bottom: 50px;
+
+  &__tabs {
+    margin-bottom: 25px;
+  }
+
+  &__tab {
+    padding: 11px 28px;
+    border-radius: 5px;
+
+    &:hover {
+      text-decoration: underline;
+    }
+
+    &.active {
+      background-color: $base-color;
+    }
+  }
+
+  &__form {
+    width: 100%;
+    position: relative;
+  }
+
+  &__input {
+    width: 100%;
+    height: 50px;
+    background-color: $base-color;
+    padding: 15px 150px 15px 20px;
+    font-size: 16px;
+    line-height: 19px;
+    border-radius: 3px;
+  }
+
+  &__submit-button {
+    height: 50px;
+    padding: 16px 38px;
+    font-weight: 700;
+    font-size: 14px;
+    line-height: 17px;
+    text-transform: uppercase;
+    letter-spacing: .12em;
+    position: absolute;
+    right: 0;
+    border-radius: 3px;
+  }
+}
