+++ src/app/user/home-page/components/tiles/tiles.component.scss
