+++ src/app/user/shared/components/carousel-dots/carousel-dots.component.scss
