+++ src/app/user/shared/components/product-carousel-menu/product-carousel-menu.component.ts
