+++ src/app/user/shared/components/product-carousel-section/product-carousel-section.component.scss
