+++ src/app/user/shared/components/product-carousel/product-carousel.component.scss
