+++ src/app/user/shared/components/product/product.component.scss
